requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# PaddleOCR và dependencies
paddleocr==2.7.3
paddlepaddle==2.5.2

# Image processing
Pillow==10.1.0
PyMuPDF==1.23.8

# Utils
python-multipart==0.0.6
